Refactor into models
This change hopefully makes working with an army and it's various properties much easier.

A context is passed in on construction and everything from then on has it and can use it, so less passing around context/data objects around.

Methods also group functionality together quite nicely and organise things, so it's easier to know where to go to for something.

It also uses svelte 5 runes so is reactive and can be used directly in the UI, but also the backend simultaneously, which is a great benefit.

There are still plenty of improvements to be made, but this is a good step towards the app being easier to maintain.