Canonical tag test 2
Previous commit (before deploy) did not fix indexing issue, so trying to completely remove the meta/canonical tags. It is hard because google is ambigious in their description of why they cannot index this page. Other pages are fine. It was also indexed a few days ago without failure so it is a recent change