Split API endpoints away from pages
This is what prevented google indexing certain pages on the site.

It turns out if you have a +server.ts for API endpoints and a +page.svelte
in the same route directory and you don't include an "Accept: text/html"
header in the GET request (to view the page) it will return "405: GET method not allowed".
You can also view this behavior yourself by curling "/armies" for example.

Google crawlers (at least when you "test" the URL for indexability) don't
include this header so they then cannot index the site. It seems like
sometimes/previously they do/did as "/armies" has been indexed in the
past, but I definitely don't want to chance it and assume anything.

You either have to split your API endpoint structure away from page files
like in this commit, or explicitly export a GET handler in your endpoints
for the edge case of for example google crawlers... but this is not
really a solution because you have to now implement the same code
SvelteKit does underneath for a "normal" GET request which isn't maintainable
or sane. SvelteKit could have at least provided an API you could
call inside your handler to revert to the default behavior, but they don't.