Fix duplicate unit/equipment/pets bug
A user found there was a bug if he removed then added back the same troop
on his existing army. This also applies to equipment and pets although
less obvious as units are just rendered as a list whereas hero stuff
is a "fixed" interface since we know that we have max 2 equipment and 1
pet per hero.

The issue was when an existing, saved unit, which has a db id is removed
and re-added on the UI, the "re-add" is just adding a whole new unit
since the old one was literally deleted from the array in-memory and we
no longer know about it. When the army was saved again, the server-side:
- didn't delete old unit because we no longer have it's existing id (which is probably correct behaviour)
- then the upsert just inserted since the id PK didn't collide since it was undefined

Deleting old units code has been cleaned up a bit and uses unitId/equipmentId/petId
which is more clear and correct I think. The upsert now will correctly updating
existing db entries even with undefined id's since there are now UNIQUE
constraints on the appropriate fields.