Improve API error handling
Several improvements have been made. Reduced indentation and clearer
reading of +server.ts http handlers. The actual action/middleware
function reads better. Form/top-level zod errors are now included in
the error response. This is particularly useful because otherwise
(and you can see this around the place) there's a pattern of doing:

`const { id } = z.object({ id: z.number() }).parse({ id: +req.params.slug })`

Just so that the errors in response included an "id" field. Now we can
make this nicer and we'll be able to still see/log the error:

`const id = z.number().parse(+req.params.slug)`

You would lose it saying "{ id: ... }" and instead it would just be
"{ form: ... }" but it doesn't matter because you have the stack in the
logs, and if you wanted to provide a more informative error if it was
coming back to the client Zod lets you customise the message.