Make server hooks file leaner
The server hooks file started getting a bit large, so I've split it out
into functions for the request handling bits, and a Server class which
handles the setup/teardown and cron job task(s).

Potentially this class could further expand and expose sub APIs which
could provide a clear place for new code and splitting of responsibilities.
Could also help with unit testing as well. How it could interact with
requests is a question to think about, so we'll see.