Move APIs into Server class
All the current army/user/notification APIs have now moved into the
Server class, and that is now available on `req.locals`. This is
admittedly quite a large commit and I have subtly improved a few bits
as well, but time constraints...

Other changes include a new DB json field parse function which gets rid
of needing ts ignore comment, and also provides a JSDoc comment on why
we have to do this, so that's useful.

Also `ArmyCtx` is quite an ambiguous type name and doesn't explain much.
It's now called `StaticGameData` which provides more insight.