Improved auth interactions
Protected auth routes go directly to google signin, skipping our own
login page. This page is now there mainly for any SEO benefit.

Added client side requireAuth method which can be used to redirect to
google signin and preserve the scroll position when coming back. This
will be useful when an action on-page requires auth but you want to
go back to the same position once the user has logged in.

Logging out will redirect you back to the page you were on.