Allow navigating to army builder page without auth
This aims to encourage the user a bit more to create an army as now the
army creation page is at least previewed behind a blurred overlay, and
not just completely blocked behind a login screen.

Also this is probably a good thing for SEO/crawlers.

At some point I may consider allowing the army creation page entirely,
but only allowing saving of it once logged in. That's a bit more advanced
and not sure if I want to do this but keeping the option there.